ai navigation fixes + dummy addition + reload animation
Ai was trying to walk to places it could not access, it would get stuck & would not take damage. I think I've fixed it for now let me know if it still happens.

dummy ai at the start that can't move and can't shoot added.

Forgot to add the reload animation for ai. It has been added and works perfectly